cleanup scheme module
