+++ plonk/code/src/scheme/mod.rs
@@ -1,0 +1,4 @@
+mod enums;
+pub mod eqns;
+
+pub use enums::*;
